unauthorization implemented

+++ app/Http/Controllers/Author/PostController.php
@@ -104,7 +104,12 @@ class PostController extends Controller
      */
     public function show(Post $post)
     {
-          return view('author.post.show',compact('post'));
+      if ($post->user_id != Auth::id())
+    {
+        Toastr::error('You are not authorized to access this post','Error');
+        return redirect()->back();
+    }
+     return view('author.post.show',compact('post'));
     }
 
     /**
@@ -115,6 +120,11 @@ class PostController extends Controller
      */
     public function edit(Post $post)
     {
+      if ($post->user_id != Auth::id())
+        {
+            Toastr::error('You are not authorized to access this post','Error');
+            return redirect()->back();
+        }
       $categories = Category::all();
       $tags = Tag::all();
       return view('author.post.edit',compact('post','categories','tags'));
@@ -129,6 +139,11 @@ class PostController extends Controller
      */
     public function update(Request $request, Post $post)
     {
+      if ($post->user_id != Auth::id())
+      {
+          Toastr::error('You are not authorized to access this post','Error');
+          return redirect()->back();
+      }
       $this->validate($request,[
          'title' => 'required',
          'image' => 'image',
@@ -180,7 +195,7 @@ class PostController extends Controller
      Toastr::success('Post Successfully Updated :)','Success');
      return redirect()->route('author.post.index');
     }
-    
+
     /**
      * Remove the specified resource from storage.
      *
@@ -189,6 +204,11 @@ class PostController extends Controller
      */
     public function destroy(Post $post)
     {
+      if ($post->user_id != Auth::id())
+      {
+          Toastr::error('You are not authorized to access this post','Error');
+          return redirect()->back();
+      }
       if (Storage::disk('public')->exists('post/'.$post->image))
       {
           Storage::disk('public')->delete('post/'.$post->image);
